Split main details template into three templates
Bridges have fewer fields compared to plain relays. It does not make
sense to have several if(!bridge) constructions in the main template, so
rename main.html to router.html and add a special bridge.html template.
As we are at it, split away error handling from template - it is now
done by a separate error.html.

Signed-off-by: Dmitry Eremin-Solenikov <dbaryshkov@gmail.com>